Fix fr truncated boot image size recorded with ElTorito boot catalog and boot info table.

+++ filesystem/iso9660/finalize.go
@@ -504,7 +504,7 @@ func (fs *FileSystem) Finalize(options FinalizeOptions) error {
 				}
 			}
 			// save the child so we can add location late
-			e.size = uint16(child.size)
+			e.size = uint32(child.size)
 			child.elToritoEntry = e
 		}
 	}
